chore(Seeder): organize seeder for test and setup
