+++ database/seeders/DefaultPermissionsSeeder.php
@@ -5,6 +5,9 @@ namespace Unusualify\Modularity\Database\Seeders;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\PermissionRegistrar;
 
 class DefaultPermissionsSeeder extends Seeder
 {
@@ -32,23 +35,66 @@ class DefaultPermissionsSeeder extends Seeder
                 'name' => 'mediaLibrary',
                 'guard_name' => 'unusual_users',
             ],
-
             ...permissionRecordsFromRoutes([
                 'User',
                 'Role',
                 'Permission',
+                'VatRate',
+                'Currency',
+                'PriceType',
                 'Payment',
-                // 'PackageContinent',
-                // 'PackageRegion',
-                // 'PackageCountry',
-                // 'PackageLanguage',
-                // 'PackageFeature',
-                // 'Package',
-                // 'VatRate',
-                // 'Currency',
-                // 'PriceType',
-                // 'Price',
             ], 'unusual_users'),
         ]);
+
+        $roleInstances = Role::all()->keyBy('name');
+
+        app(PermissionRegistrar::class)->forgetCachedPermissions();
+
+        // Assign permissions to roles
+        // Admin gets most permissions except some sensitive ones
+        $roleInstances['admin']->givePermissionTo(Permission::all()->except([
+            'user_forceDelete', 'user_bulkForceDelete',
+            'role_forceDelete', 'role_bulkForceDelete',
+            'permission_forceDelete', 'permission_bulkForceDelete',
+            'vat-rate_forceDelete', 'vat_rate_bulkForceDelete',
+            'currency_forceDelete', 'currency_bulkForceDelete',
+            'price-type_forceDelete', 'price-type_bulkForceDelete',
+            'payment_forceDelete', 'payment_bulkForceDelete',
+        ]));
+
+        // Manager gets operational permissions
+        $roleInstances['manager']->givePermissionTo([
+            'dashboard', 'mediaLibrary',
+            'user_view', 'user_create', 'user_edit',
+            'vat-rate_view', 'vat-rate_create', 'vat-rate_edit',
+            'currency_view', 'currency_create', 'currency_edit',
+            'price-type_view', 'price-type_create', 'price-type_edit',
+            'payment_view', 'payment_create', 'payment_edit',
+        ]);
+
+        // Editor gets content-related permissions
+        $roleInstances['editor']->givePermissionTo([
+            'dashboard', 'mediaLibrary',
+        ]);
+
+        // Reporter gets view-only permissions
+        $roleInstances['reporter']->givePermissionTo([
+            'dashboard', 'mediaLibrary',
+        ]);
+
+        // Client Manager gets client-specific permissions
+        $roleInstances['client-manager']->givePermissionTo([
+            'dashboard',
+            'user_view', 'user_create', 'user_edit', 'user_delete',
+            'payment_view', 'payment_create',
+        ]);
+
+        // Client Assistant gets limited client permissions
+        $roleInstances['client-assistant']->givePermissionTo([
+            'dashboard',
+            'user_view',
+            'payment_view',
+        ]);
+
     }
 }
